[CBR-374] Aeson/Arbitrary instances for Kernel/WL types
- Many Kernel and WalletLayer types used in WalletLayer errors have
Aeson/Arbitrary instances with this commit. While there seems to be no
absolute certainty of what the V2 API error handling should be, this is
a preliminary first step.
- Existing `Arbitrary` instances with `oneof []` are expanded
- The WalletLayer `Arbitrary` instances still in the WalletLayer module
are moved to Cardano.Wallet.WalletLayer.QuickCheck where they belong.
- `aesonRoundtripProp` for HdWallet types (discovered a problem with the
Arbitrary instances when roundtrip testing larger error types. This has
not ended up in this PR, though the problem was fixed separately.)

[CBR-374] `aesonRoundtripProp ` for HdWallet types

Ensuring `aesonRoundtripProp ` on these types is a first step of testing roundtrips on larger error types.